Notfound-page translated, InternalLink>edit bug fix, FileTools: Delete file with Exceptions

+++ pwTools/file/FileTools.php
@@ -49,6 +49,60 @@ class FileTools {
 		}
 	}
 	
+	public static function getUnixFilePermission($filename) {
+		if (strlen($filename) == 0) {
+			throw new Exception("File strings cannot be empty");
+		}
+		if (!is_file($filename)) {
+			throw new Exception("File '$filename' does not exist!");
+		}
+		
+		$perms = fileperms($filename);
+		
+		if (($perms & 0xC000) == 0xC000) {
+			// Socket
+			$info = 's';
+		} elseif (($perms & 0xA000) == 0xA000) {
+			// Symbolic Link
+			$info = 'l';
+		} elseif (($perms & 0x8000) == 0x8000) {
+			// Regular
+			$info = '-';
+		} elseif (($perms & 0x6000) == 0x6000) {
+			// Block special
+			$info = 'b';
+		} elseif (($perms & 0x4000) == 0x4000) {
+			// Directory
+			$info = 'd';
+		} elseif (($perms & 0x2000) == 0x2000) {
+			// Character special
+			$info = 'c';
+		} elseif (($perms & 0x1000) == 0x1000) {
+			// FIFO pipe
+			$info = 'p';
+		} else {
+			// Unknown
+			$info = 'u';
+		}
+		
+		// Owner
+		$info .= (($perms & 0x0100) ? 'r' : '-');
+		$info .= (($perms & 0x0080) ? 'w' : '-');
+		$info .= (($perms & 0x0040) ? (($perms & 0x0800) ? 's' : 'x' ) : (($perms & 0x0800) ? 'S' : '-'));
+		
+		// Group
+		$info .= (($perms & 0x0020) ? 'r' : '-');
+		$info .= (($perms & 0x0010) ? 'w' : '-');
+		$info .= (($perms & 0x0008) ? (($perms & 0x0400) ? 's' : 'x' ) : (($perms & 0x0400) ? 'S' : '-'));
+		
+		// World
+		$info .= (($perms & 0x0004) ? 'r' : '-');
+		$info .= (($perms & 0x0002) ? 'w' : '-');
+		$info .= (($perms & 0x0001) ? (($perms & 0x0200) ? 't' : 'x' ) : (($perms & 0x0200) ? 'T' : '-'));
+		
+		return $info;
+	}
+	
 	private static function removeDirectoryRec($dir) {
 		if (is_dir($dir)) {
 			$objects = scandir($dir);
@@ -57,6 +111,9 @@ class FileTools {
 					if (filetype($dir."/".$object) == "dir") {
 						self::removeDirectoryRec($dir."/".$object);
 					} else {
+						if(!is_writeable($dir."/".$object)) {
+							throw new Exception("Your are not allowed to delete:<br />'$dir'!<br />Permissions are ".self::getUnixFilePermission($dir."/".$object));
+						}
 						if (!unlink($dir."/".$object)) {
 							throw new Exception("Unable to remove directory '".$dir."/".$object."'");
 						}
@@ -75,6 +132,21 @@ class FileTools {
 			throw new Exception("Unable to find directory '".$dir."'");
 		}
 		self::removeDirectoryRec($dir);
+	}
+	
+	public static function removeFile($filename) {
+		if (strlen($filename) == 0) {
+			throw new Exception("File strings cannot be empty");
+		}
+		if (!is_file($filename)) {
+			throw new Exception("The file '$filename' does not exist!");
+		}
+		if(!is_writeable($filename)) {
+			throw new Exception("Your are not allowed to delete:<br />'$filename'!<br />Permissions are ".self::getUnixFilePermission($filename));
+		}
+		if (!unlink($filename)) {
+			throw new Exception("Unable to remove file '$filename'");
+		}
 	}
 	
 	public static function getTextFileFormat($text) {
